requirements.txt
streamlit
opencv-python
dlib
numpy
pillow
tensorflow